feat: Refactor toast component structure and improve type definitions
